// src/server/gen/ris-journeys/types.gen.ts
export type CodeShare = {
    /**
     * Code of cooperating airline [IATA-Code der Fluggesellschaft].
     */
    airlineCode: string;
    /**
     * Flightnumber of cooperating airline journey [Flugnummer des Fluges der kooperierenden Fluggesellschaft].
     */
    flightnumber: number;
};
// ...
export type EventType = 'ARRIVAL' | 'DEPARTURE';
// ...
export type JourneyEvent = {
    /**
     * Indicates whether event is additional, meaning not be part of the regular schedule [Zusatzhalt]. Note: In case an event is cancelled and additional, it's considered as an cancelled additional stop [zurückgenommener Zusatzhalt].
     */
    additional?: boolean;
    /**
     * ID of arrival or departure, depends on event type.
     */
    arrivalOrDepartureID: string;
    /**
     * Indicates whether event is cancelled [Haltausfall]. Note: In case an event is cancelled and additional, it's considered as an cancelled additional stop [zurückgenommener Zusatzhalt].
     */
    cancelled?: boolean;
    /**
     * List of codeshares [Code-Teilungen mit Flügen verschiedener Fluggesellschaften] for this particular journey at this arrival / departure.
     */
    codeshares?: Array<CodeShare>;
    /**
     * List of message-ids [Nachrichten IDs] for this particular event. IDs are pointing to the global 'messages' structure, containing disruption-communications [Störungskommunikationen], attributes [Fahrtmerkmale], notes [Hinweise], ris cause codes [RIS Kundengründe] and ris quality deviations [RIS Qualitätsabweichungen].
     */
    messages?: Array<number>;
    /**
     * Indicates whether passengers are not allowed to enter / leave [kein Fahrgastwechsel].
     */
    noPassengerChange?: boolean;
    /**
     * Indicates whether arrival / departure is an on demand stop [Bedarfshalt].
     */
    onDemand?: boolean;
    /**
     * Actual platform [Gleis, Bahnsteig, Plattform] the transport arrives / departs at, may be empty.
     */
    platform?: string;
    /**
     * Ifopt [DHID] of actual platform [Gleis, Bahnsteig, Plattform] the transport arrives / departs at, may be empty.
     */
    platformIfopt?: string;
    /**
     * Scheduled platform [Gleis, Bahnsteig, Plattform] the transport arrives / departs at.
     */
    platformSchedule?: string;
    /**
     * Ifopt [DHID] of scheduled platform [Gleis, Bahnsteig, Plattform] the transport arrives / departs at.
     */
    platformScheduleIfopt?: string;
    /**
     * List of transports this journey at this particular event is reliefed by [Entlastungszug].
     */
    reliefBy?: Array<TransportDestinationRef>;
    /**
     * List of transports this journey at this particular event reliefs for [Entlastungszug].
     */
    reliefFor?: Array<TransportDestinationRef>;
    /**
     * List of transports this journey at this particular event is replaced by [Ersatzzug].
     */
    replacedBy?: Array<TransportDestinationRef>;
    /**
     * List of transports this journey at this particular event replaces [Ersatzzug].
     */
    replacementFor?: Array<TransportDestinationRef>;
    stopPlace: StopPlaceInJourney;
    /**
     * Best known time information of stop as fully-qualified-date (ISO-8601 with time-zone or offset).
     */
    time: string;
    /**
     * Scheduled time [Soll] of stop as fully-qualified-date (ISO-8601 with time-zone or offset).
     */
    timeSchedule: string;
    /**
     * Specifies on which information 'time' is based.
     * - SCHEDULE (Time source is schedule [Plan / Soll])
     * - PREVIEW (Time source is preview / forecast [Vorschau / Disposition / Prognose])
     * - REAL (Time source is real [Echt = passiert (kann nur in die Vergangenheit gesetzt werden)
     */
    timeType: string;
    transport: TransportWithDirection;
    /**
     * List of journeys this journey at this particular event travels with [Vereinigung].
     */
    travelsWith?: Array<TransportDestinationPortionWorkingRef>;
    type: EventType;
    /**
     * Indicates the presence of a vehicle-sequence [Wagenreihung / Formation] for this journey at this particular arrival / departure.
     */
    vehicleSequenceAvailable?: boolean;
};
// ...
export type StopPlaceDifferingInJourney = {
    /**
     * Eva number of stop-place [Haltestelle].
     */
    evaNumber: string;
    /**
     * Transmodel identifier for fixed objects, in germany DHID = Deutschlandweite Halt ID also known as global id, of stop-place [Haltestelle].
     */
    ifopt?: string;
    /**
     * Name for stop-place [Haltestelle] in fixed language 'DE'.
     */
    name: string;
};

/**
 * Comprehensive stop-place [Haltestelle] information.
 */
export type StopPlaceEmbedded = {
    /**
     * Eva number of stop-place [Haltestelle].
     */
    evaNumber: string;
    /**
     * Transmodel identifier for fixed objects, in germany DHID = Deutschlandweite Halt ID also known as global id, of stop-place [Haltestelle].
     */
    ifopt?: string;
    /**
     * Name for stop-place [Haltestelle] in fixed language 'DE'.
     */
    name: string;
};

/**
 * Comprehensive stop-place [Haltestelle] information.
 */
export type StopPlaceEmbeddedWithCancel = {
    /**
     * Indicates whether the stop ie departure / arrival has been cancelled [Haltausfall].
     */
    cancelled?: boolean;
    /**
     * Eva number of stop-place [Haltestelle].
     */
    evaNumber: string;
    /**
     * Transmodel identifier for fixed objects, in germany DHID = Deutschlandweite Halt ID also known as global id, of stop-place [Haltestelle].
     */
    ifopt?: string;
    /**
     * Name for stop-place [Haltestelle] in fixed language 'DE'.
     */
    name: string;
};

/**
 * Stop-place [Haltestelle] information for a stop-place a transport [Verkehrsart / Verkehrsmittel] departs / arrives at.
 */
export type StopPlaceInJourney = {
    differingStopPlace?: StopPlaceDifferingInJourney;
    /**
     * Eva number of stop-place [Haltestelle].
     */
    evaNumber: string;
    /**
     * Transmodel identifier for fixed objects, in germany DHID = Deutschlandweite Halt ID also known as global id, of stop-place [Haltestelle].
     */
    ifopt?: string;
    /**
     * Name for stop-place [Haltestelle] in fixed language 'DE'.
     */
    name: string;
};
// ...
export type TransportDestinationPortionWorkingRef = {
    /**
     * Category of the transport [externe Fahrtgattung].
     */
    category: string;
    /**
     * Internal category of the transport [interne Fahrtgattung vor Ausgabensteuerung). Must not be shown to customers.
     */
    categoryInternal: string;
    destination: StopPlaceEmbeddedWithCancel;
    /**
     * Deutschlandweite Fahrt ID (DFID).
     */
    dfid?: string;
    differingDestination?: StopPlaceEmbedded;
    /**
     * Deutschlandweite Teil-Linien ID (DTID).
     */
    dtid?: string;
    /**
     * Description of the journey [Fahrtbezeichnung] that must be used to inform the traveller.
     */
    journeyDescription: string;
    /**
     * ID of journey [FahrtID].
     */
    journeyID: string;
    /**
     * Number of the transport [Fahrtnummer].
     */
    journeyNumber: number;
    /**
     * Marketing or product name of the transport, for instance 'Sprinter' or 'Schwarzwaldexpress' etc. [Marketing Name / Produkt Name].
     */
    label?: string;
    /**
     * Line of the transport [Linie].
     */
    line?: string;
    /**
     * Operational journey-number [betriebliche Fahrtnummer].
     */
    operationalJourneyNumber: string;
    separationAt?: StopPlaceEmbedded;
    /**
     * Type of transport. Possible values are:
     * - HIGH_SPEED_TRAIN (High speed train [Hochgeschwindigkeitszug])
     * - INTERCITY_TRAIN (Inter city train [Intercityzug])
     * - INTER_REGIONAL_TRAIN (Inter regional train [Interregiozug])
     * - REGIONAL_TRAIN (Regional train [Regionalzug])
     * - CITY_TRAIN (City train [S-Bahn])
     * - SUBWAY (Subway [U-Bahn])
     * - TRAM (Tram [Strassenbahn])
     * - BUS (Bus [Bus])
     * - FERRY (Ferry [Faehre])
     * - SHUTTLE (Shuttle [Ruftaxi])
     * - FLIGHT (Flight [Flug])
     * - CHARTER_TRAIN (Charter train [Charterzug])
     * - UNKNOWN (Unknown)
     */
    type: string;
};

/**
 * Transport [Verkehrsart / Verkehrsmittel] information for journey references with destination.
 */
export type TransportDestinationRef = {
    /**
     * Category of the transport [externe Fahrtgattung].
     */
    category: string;
    /**
     * Internal category of the transport [interne Fahrtgattung vor Ausgabensteuerung). Must not be shown to customers.
     */
    categoryInternal: string;
    destination: StopPlaceEmbeddedWithCancel;
    /**
     * Deutschlandweite Fahrt ID (DFID).
     */
    dfid?: string;
    differingDestination?: StopPlaceEmbedded;
    /**
     * Deutschlandweite Teil-Linien ID (DTID).
     */
    dtid?: string;
    /**
     * Description of the journey [Fahrtbezeichnung] that must be used to inform the traveller.
     */
    journeyDescription: string;
    /**
     * ID of journey [FahrtID].
     */
    journeyID: string;
    /**
     * Number of the transport [Fahrtnummer].
     */
    journeyNumber: number;
    /**
     * Marketing or product name of the transport, for instance 'Sprinter' or 'Schwarzwaldexpress' etc. [Marketing Name / Produkt Name].
     */
    label?: string;
    /**
     * Line of the transport [Linie].
     */
    line?: string;
    /**
     * Operational journey-number [betriebliche Fahrtnummer].
     */
    operationalJourneyNumber: string;
    /**
     * Type of transport. Possible values are:
     * - HIGH_SPEED_TRAIN (High speed train [Hochgeschwindigkeitszug])
     * - INTERCITY_TRAIN (Inter city train [Intercityzug])
     * - INTER_REGIONAL_TRAIN (Inter regional train [Interregiozug])
     * - REGIONAL_TRAIN (Regional train [Regionalzug])
     * - CITY_TRAIN (City train [S-Bahn])
     * - SUBWAY (Subway [U-Bahn])
     * - TRAM (Tram [Strassenbahn])
     * - BUS (Bus [Bus])
     * - FERRY (Ferry [Faehre])
     * - SHUTTLE (Shuttle [Ruftaxi])
     * - FLIGHT (Flight [Flug])
     * - CHARTER_TRAIN (Charter train [Charterzug])
     * - UNKNOWN (Unknown)
     */
    type: string;
};
// ...
export type TransportWithDirection = {
    administration: Administration;
    /**
     * Category of the transport [externe Fahrtgattung].
     */
    category: string;
    /**
     * Internal category of the transport [interne Fahrtgattung vor Ausgabensteuerung). Must not be shown to customers.
     */
    categoryInternal: string;
    /**
     * Deutschlandweite Fahrt ID (DFID).
     */
    dfid?: string;
    direction?: DirectionInfo;
    /**
     * Deutschlandweite Teil-Linien ID (DTID).
     */
    dtid?: string;
    /**
     * Description of the journey [Fahrtbezeichnung] that must be used to inform the traveller.
     */
    journeyDescription: string;
    /**
     * Number of the transport [Fahrtnummer].
     */
    journeyNumber: number;
    /**
     * Marketing or product name of the transport, for instance 'Sprinter' or 'Schwarzwaldexpress' etc. [Marketing Name / Produkt Name].
     */
    label?: string;
    /**
     * Line of the transport [Linie].
     */
    line?: string;
    /**
     * Operational journey-number [betriebliche Fahrtnummer].
     */
    operationalJourneyNumber: string;
    replacementTransport?: ReplacementTransport;
    /**
     * Type of transport. Possible values are:
     * - HIGH_SPEED_TRAIN (High speed train [Hochgeschwindigkeitszug])
     * - INTERCITY_TRAIN (Inter city train [Intercityzug])
     * - INTER_REGIONAL_TRAIN (Inter regional train [Interregiozug])
     * - REGIONAL_TRAIN (Regional train [Regionalzug])
     * - CITY_TRAIN (City train [S-Bahn])
     * - SUBWAY (Subway [U-Bahn])
     * - TRAM (Tram [Strassenbahn])
     * - BUS (Bus [Bus])
     * - FERRY (Ferry [Faehre])
     * - SHUTTLE (Shuttle [Ruftaxi])
     * - FLIGHT (Flight [Flug])
     * - CHARTER_TRAIN (Charter train [Charterzug])
     * - UNKNOWN (Unknown)
     */
    type: string;
};